created validator.go and moved GetVal/AddVal/RemoveVal
